Move tests to proper location (they are not integration tests)
